requirements.txt
# Core dependencies
flask==2.3.3
langchain==0.1.0
langchain-community==0.0.20

# Vector store
chromadb==0.4.22
faiss-cpu==1.7.4

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
python-multipart==0.0.6

# LLM providers
google-generativeai==0.3.2
requests==2.31.0

# Embeddings - Fixed versions for compatibility
sentence-transformers>=2.2.2
huggingface-hub==0.19.4
transformers>=4.30.0

# Utilities
python-dotenv==1.0.0
werkzeug==2.3.7 